+++ Firmware/Application.X/modes/keys/keys.c
@@ -79,7 +79,7 @@ void keys_create_next(void) {
     }
 
     mode_data.keys.seconds = key_at % 60;
-    mode_data.keys.minutes = (key_at - mode_data.keys.seconds) / 60;
+    mode_data.keys.minutes = (uint8_t) ((key_at - mode_data.keys.seconds) / 60);
     mode_data.keys.key = rng_generate8(&game.module_seed, KEYS_RNG_MASK) % 3;
 
     uint8_t minutes = mode_data.keys.minutes % 10;
